backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-environ==0.12.0
gunicorn==23.0.0
psycopg-binary==3.2.10
pymongo==4.15.2
pydantic[email]==2.11.10
whitenoise==6.11.0
